+++ packages/workers-utils/src/config/index.ts
@@ -44,9 +44,11 @@ export {
 	CONTAINER_IMAGES_BINDING,
 	getContainerDurableObjectClassNames,
 	getContainerNameToClassNameMap,
+	getDurableObjectClassNameToUseSQLiteMap,
 	getDurableObjectContainerApps,
 	isDurableObjectContainerApp,
 	resolveContainerClassName,
+	validateDurableObjectContainerApplications,
 } from "./containers";
 export type { DurableObjectContainerApp } from "./containers";
