+++ packages/workers-utils/src/index.ts
@@ -17,9 +17,11 @@ export {
 	CONTAINER_IMAGES_BINDING,
 	getContainerDurableObjectClassNames,
 	getContainerNameToClassNameMap,
+	getDurableObjectClassNameToUseSQLiteMap,
 	getDurableObjectContainerApps,
 	isDurableObjectContainerApp,
 	resolveContainerClassName,
+	validateDurableObjectContainerApplications,
 } from "./config/containers";
 export type { DurableObjectContainerApp } from "./config/containers";
 export {
