+++ packages/wrangler/src/containers/config.ts
@@ -11,6 +11,7 @@ import {
 	isDurableObjectContainerApp,
 	resolveContainerClassName,
 	UserError,
+	validateDurableObjectContainerApplications,
 } from "@cloudflare/workers-utils";
 import { getDurableObjectClassNameToUseSQLiteMap } from "../dev/class-names-sqlite";
 import { getOrSelectAccountId } from "../user";
@@ -44,49 +45,6 @@ function isRootObservabilityLogsEnabled(
 		observability?.logs?.enabled === true ||
 		(observability?.enabled === true && observability?.logs?.enabled !== false)
 	);
-}
-
-/**
- * Validate that every Durable Object-managed container belongs to this Worker.
- */
-export function validateDurableObjectContainerApplications(
-	config: Config
-): void {
-	const allDOs = getDurableObjectClassNameToUseSQLiteMap(
-		config.migrations,
-		config.exports
-	);
-
-	for (const container of getDurableObjectContainerApps(config.containers)) {
-		const maybeBoundDO = config.durable_objects.bindings.find(
-			(durableObject) => durableObject.class_name === container.class_name
-		);
-		const useSQLite = allDOs.get(container.class_name);
-		if (useSQLite === undefined && maybeBoundDO === undefined) {
-			throw new UserError(
-				`The container class_name ${container.class_name} does not match any durable object class_name defined in your Wrangler config file. Note that the durable object must be defined in the same script as the container.`,
-				{ telemetryMessage: "no DO defined that matches container class_name" }
-			);
-		}
-		if (maybeBoundDO?.script_name !== undefined) {
-			throw new UserError(
-				`The container ${container.name} is referencing the durable object ${container.class_name}, which appears to be defined on the ${maybeBoundDO.script_name} Worker instead (via the 'script_name' field). You cannot configure a container on a Durable Object that is defined in another Worker.`,
-				{
-					telemetryMessage:
-						"container class_name refers to an external durable object",
-				}
-			);
-		}
-		if (useSQLite === false) {
-			throw new UserError(
-				`The container ${container.name} references Durable Object class ${container.class_name}, which uses the legacy KV storage backend. Durable Object-managed Containers require SQLite-backed Durable Objects.`,
-				{
-					telemetryMessage:
-						"durable object container class uses legacy storage",
-				}
-			);
-		}
-	}
 }
 
 /**
